Refactor shared book helpers
